Restructure: extract algorithm internals into algorithm/ directory
