+++ raptor/src/lib.rs
@@ -143,13 +143,19 @@ use std::collections::BinaryHeap;
 use std::sync::Mutex;
 
 use fixedbitset::FixedBitSet;
-use smallvec::SmallVec;
 
+mod algorithm;
 mod endpoints;
 mod ids;
 mod journey;
 mod label;
 mod time;
+
+use crate::algorithm::boarding::BoardingTree;
+use crate::algorithm::label_bag::LabelBag;
+use crate::algorithm::per_call::run_per_call_query;
+use crate::algorithm::range::filter_range_pareto_front;
+use crate::algorithm::range::raptor_range_rrap_arrival;
 
 pub mod gtfs;
 pub mod labels;
@@ -178,660 +184,6 @@ mod test;
 /// Internal round-counter type, used only for indexing label arrays.
 /// User-facing transfer caps are passed as [`Transfers`].
 pub(crate) type K = usize;
-
-/// One reconstructable step in a journey: either a transit boarding event
-/// (route-scan) or a walk along a footpath. Walks do not consume a round —
-/// they happen *within* round `k` at the stop they alight on — so the
-/// reconstruction logic chains through walk entries without decrementing
-/// the round index.
-///
-/// `parent_arrival` disambiguates which Pareto-optimal label at the
-/// parent stop (in the parent stop's bag) was extended to produce
-/// the label this step belongs to. For single-criterion `ArrivalTime`
-/// the parent bag is size-1, so this field is redundant; for
-/// multi-criterion impls it lets reconstruction follow the right
-/// label through the bag.
-#[derive(Debug, Clone, Copy)]
-enum Step {
-    Boarded {
-        from: StopIdx,
-        route: RouteIdx,
-        parent_arrival: SecondOfDay,
-    },
-    Walked {
-        from: StopIdx,
-        parent_arrival: SecondOfDay,
-    },
-}
-
-/// Boarding tree key: `(round, stop, label_arrival)`. The third
-/// component disambiguates Pareto-optimal labels with distinct
-/// arrival times in the same `(round, stop)` bag.
-type BoardingTree = BTreeMap<(K, StopIdx, SecondOfDay), Step>;
-
-/// A Pareto front of [`Label`]s at a single `(round, stop)` cell.
-/// Backed by `SmallVec<[L; 8]>` — for single-criterion `ArrivalTime`
-/// the bag is always size 1 and stays inline; for multi-criterion
-/// impls it grows up to 8 inline before spilling.
-#[derive(Debug, Clone)]
-struct LabelBag<L: Label> {
-    items: SmallVec<[L; 8]>,
-}
-
-impl<L: Label> LabelBag<L> {
-    fn new() -> Self {
-        Self {
-            items: SmallVec::new(),
-        }
-    }
-
-    fn is_empty(&self) -> bool {
-        self.items.is_empty()
-    }
-
-    fn iter(&self) -> impl Iterator<Item = &L> {
-        self.items.iter()
-    }
-
-    /// Try to insert `new`. Returns `true` if added (and removes any
-    /// items it strictly dominates); returns `false` if some existing
-    /// item weakly dominates `new` (no change).
-    fn insert(&mut self, new: L) -> bool {
-        for item in &self.items {
-            if item.dominates(&new) {
-                return false;
-            }
-        }
-        self.items.retain(|item| !new.dominates(item));
-        self.items.push(new);
-        true
-    }
-
-    /// Minimum `arrival()` across the bag, or `SecondOfDay::MAX` if empty.
-    fn min_arrival(&self) -> SecondOfDay {
-        self.items
-            .iter()
-            .map(|l| l.arrival())
-            .min()
-            .unwrap_or(SecondOfDay::MAX)
-    }
-
-    fn clear(&mut self) {
-        self.items.clear();
-    }
-}
-
-impl<L: Label> Default for LabelBag<L> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Relax footpaths from every stop in `sources` at round `k`, propagating
-/// to a fixed point.
-///
-/// Iteratively improves `labels[k][p_dash]` and the τ\* table
-/// (`best_arrival`) until no walk produces a strictly better arrival.
-/// Each improvement records a `Walked` step in the boarding tree so that
-/// `reconstruct_journey` can later trace back through the walk leg.
-///
-/// Because relaxation iterates to a fixed point, the trait's footpath
-/// relation does **not** need to be transitively closed: if A→B and B→C
-/// are both in the relation, the algorithm chains them within a single
-/// round and reaches C with the combined walk time.
-///
-/// Stops that should be added to the marked set for the next round are
-/// pushed onto `out`, gated by `pt_threshold` (the current best effective
-/// arrival at any target). Caller drains `out` between calls.
-/// Single-pass relaxation for adapters that report `true` from
-/// [`Timetable::footpaths_are_transitively_closed`]. One walk per
-/// source — chained walks are unnecessary because every reachable
-/// pair is already a direct edge in the relation.
-///
-/// `O(E)` per round, no heap. Always sound when the closure
-/// precondition holds; produces the same labels and boarding tree as
-/// [`relax_footpaths_round`] in that case.
-/// Try to insert `(label, step)` into `labels[k][stop]` and the
-/// boarding tree. Updates `best_arrival[stop]` and marks the stop
-/// in `out` if the new label improves on `pt_threshold`. Returns
-/// `true` if any insertion happened.
-#[allow(clippy::too_many_arguments)]
-fn insert_into_bag<L: Label>(
-    labels: &mut [Vec<LabelBag<L>>],
-    best_arrival: &mut [LabelBag<L>],
-    board_detail: &mut BoardingTree,
-    out: &mut Vec<StopIdx>,
-    ever_reached: &mut FixedBitSet,
-    pt_threshold: SecondOfDay,
-    k: K,
-    stop: StopIdx,
-    label: L,
-    step: Step,
-) -> bool {
-    let added = labels[k][stop.idx()].insert(label);
-    if !added {
-        return false;
-    }
-    board_detail.insert((k, stop, label.arrival()), step);
-    best_arrival[stop.idx()].insert(label);
-    ever_reached.insert(stop.idx());
-    if label.arrival() < pt_threshold {
-        out.push(stop);
-    }
-    true
-}
-
-#[allow(clippy::too_many_arguments)]
-fn relax_footpaths_round_closed<T: Timetable + ?Sized, L: Label>(
-    timetable: &T,
-    k: K,
-    labels: &mut [Vec<LabelBag<L>>],
-    best_arrival: &mut [LabelBag<L>],
-    board_detail: &mut BoardingTree,
-    sources: &FixedBitSet,
-    pt_threshold: SecondOfDay,
-    out: &mut Vec<StopIdx>,
-    ever_reached: &mut FixedBitSet,
-) {
-    // Walk each source bag's labels once. With closure, chained walks
-    // are unnecessary because every reachable pair is already a direct
-    // edge in the relation.
-    let mut staged: SmallVec<[L; 8]> = SmallVec::new();
-    for stop_bit in sources.ones() {
-        let stop = StopIdx::new(stop_bit as u32);
-        if labels[k][stop.idx()].is_empty() {
-            continue;
-        }
-        // Snapshot the source bag so we don't iterate while mutating
-        // labels[k] (which can include the source itself).
-        staged.clear();
-        staged.extend(labels[k][stop.idx()].iter().copied());
-
-        for &p_dash in timetable.get_footpaths_from(stop) {
-            let walk = timetable.get_transfer_time(stop, p_dash);
-            for source_label in &staged {
-                let via_walk = source_label.extend_by_footpath(walk);
-                insert_into_bag(
-                    labels,
-                    best_arrival,
-                    board_detail,
-                    out,
-                    ever_reached,
-                    pt_threshold,
-                    k,
-                    p_dash,
-                    via_walk,
-                    Step::Walked {
-                        from: stop,
-                        parent_arrival: source_label.arrival(),
-                    },
-                );
-            }
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn relax_footpaths_round<T: Timetable + ?Sized, L: Label>(
-    timetable: &T,
-    k: K,
-    labels: &mut [Vec<LabelBag<L>>],
-    best_arrival: &mut [LabelBag<L>],
-    board_detail: &mut BoardingTree,
-    sources: &FixedBitSet,
-    pt_threshold: SecondOfDay,
-    out: &mut Vec<StopIdx>,
-    heap: &mut BinaryHeap<Reverse<(SecondOfDay, u32)>>,
-    ever_reached: &mut FixedBitSet,
-) {
-    // Multi-source Dijkstra over the footpath graph at round `k`,
-    // ordered by `LabelBag::min_arrival()`. Each source's initial
-    // priority is its current bag's minimum arrival; transfer times
-    // are non-negative so Dijkstra is sound. Uses lazy deletion —
-    // stale heap entries are skipped on pop.
-    //
-    // For multi-criterion bags this propagates only the min-arrival
-    // label per stop; non-min labels in a source bag don't drive
-    // further relaxation. This is sound for arrival-time pruning but
-    // may miss Pareto-optimal walk chains where a non-min-arrival
-    // label dominates downstream — a v0.12 concern.
-    heap.clear();
-    for bit in sources.ones() {
-        let min_arr = labels[k][bit].min_arrival();
-        if min_arr != SecondOfDay::MAX {
-            heap.push(Reverse((min_arr, bit as u32)));
-        }
-    }
-
-    let mut staged: SmallVec<[L; 8]> = SmallVec::new();
-    while let Some(Reverse((arrival, stop_bit))) = heap.pop() {
-        let stop = StopIdx::new(stop_bit);
-        let stop_min = labels[k][stop.idx()].min_arrival();
-        // Skip stale entries — a strictly better label was popped earlier.
-        if arrival > stop_min {
-            continue;
-        }
-        staged.clear();
-        staged.extend(labels[k][stop.idx()].iter().copied());
-
-        for &p_dash in timetable.get_footpaths_from(stop) {
-            let walk = timetable.get_transfer_time(stop, p_dash);
-            let mut any_added = false;
-            for source_label in &staged {
-                let via_walk = source_label.extend_by_footpath(walk);
-                if insert_into_bag(
-                    labels,
-                    best_arrival,
-                    board_detail,
-                    out,
-                    ever_reached,
-                    pt_threshold,
-                    k,
-                    p_dash,
-                    via_walk,
-                    Step::Walked {
-                        from: stop,
-                        parent_arrival: source_label.arrival(),
-                    },
-                ) {
-                    any_added = true;
-                }
-            }
-            if any_added {
-                heap.push(Reverse((
-                    labels[k][p_dash.idx()].min_arrival(),
-                    p_dash.get(),
-                )));
-            }
-        }
-    }
-}
-
-/// Returns the minimum of `best_arrival[t].min_arrival() + w` across
-/// all `(t, w)` in `targets`, saturating on overflow. Returns
-/// `SecondOfDay::MAX` if every target is unreached.
-fn best_to_any_target<L: Label>(
-    best_arrival: &[LabelBag<L>],
-    targets: &[(StopIdx, Duration)],
-) -> SecondOfDay {
-    targets
-        .iter()
-        .map(|&(t, w)| best_arrival[t.idx()].min_arrival() + w)
-        .min()
-        .unwrap_or(SecondOfDay::MAX)
-}
-
-/// Reconstruct a single candidate plan terminating at the target
-/// label `(pt, target_arrival)` at round `k`. Traces back through
-/// the boarding tree (which is keyed on `(round, stop, label_arrival)`
-/// to disambiguate Pareto-optimal labels in the same bag). Returns
-/// `Some((origin, plan))` if the trace reaches some stop in `origins`,
-/// `None` otherwise.
-fn reconstruct_journey(
-    tree: &BoardingTree,
-    origins: &FixedBitSet,
-    pt: StopIdx,
-    target_arrival: SecondOfDay,
-    k: K,
-) -> Option<(StopIdx, Vec<(RouteIdx, StopIdx)>)> {
-    if tree.is_empty() {
-        return None;
-    }
-
-    let mut plan = Vec::with_capacity(k);
-    let mut parent = pt;
-    let mut parent_arrival = target_arrival;
-    let mut inner_k = k;
-    // Defensive bound to avoid pathological loops; 100 walk-hops per
-    // round is well beyond anything realistic.
-    let mut budget = (k + 1) * 100;
-
-    while !origins.contains(parent.idx()) && budget > 0 {
-        budget -= 1;
-
-        let Some(step) = tree.get(&(inner_k, parent, parent_arrival)).copied() else {
-            break;
-        };
-
-        match step {
-            Step::Boarded {
-                from,
-                route,
-                parent_arrival: pa,
-            } => {
-                plan.push((route, parent));
-                parent = from;
-                parent_arrival = pa;
-                if inner_k == 0 {
-                    break;
-                }
-                inner_k -= 1;
-            }
-            Step::Walked {
-                from,
-                parent_arrival: pa,
-            } => {
-                parent = from;
-                parent_arrival = pa;
-                // walks do not consume a round
-            }
-        }
-    }
-
-    if !plan.is_empty() && origins.contains(parent.idx()) {
-        plan.reverse();
-        Some((parent, plan))
-    } else {
-        None
-    }
-}
-
-/// Reconstruct one Pareto-front-of-trip-counts worth of journeys
-/// for the given targets at the current cache state. Used by both
-/// the per-call algorithm (after running rounds) and the rRAPTOR
-/// scan (snapshot at end of each τ scan).
-///
-/// Walks each target × round × label-bag entry, traces back via
-/// `reconstruct_journey`, and applies the target's walk-time offset.
-/// Output is unfiltered (caller applies any Pareto front filtering).
-fn extract_target_journeys<L: Label>(
-    labels: &[Vec<LabelBag<L>>],
-    board_detail: &BoardingTree,
-    origin_set: &FixedBitSet,
-    targets: &[(StopIdx, Duration)],
-    transfers: usize,
-) -> Vec<Journey<L>> {
-    let mut journeys: Vec<Journey<L>> = Vec::new();
-    for &(target, walk) in targets {
-        #[allow(clippy::needless_range_loop)]
-        for k in 1..=transfers {
-            for raw_label in labels[k][target.idx()].iter() {
-                let raw_arr = raw_label.arrival();
-                if raw_arr == SecondOfDay::MAX {
-                    continue;
-                }
-                let Some((origin, plan)) =
-                    reconstruct_journey(board_detail, origin_set, target, raw_arr, k)
-                else {
-                    continue;
-                };
-                let label = raw_label.extend_by_footpath(walk);
-                journeys.push(Journey {
-                    origin,
-                    target,
-                    plan,
-                    label,
-                });
-            }
-        }
-    }
-    journeys
-}
-
-/// Run round-0 footpath relaxation followed by rounds `1..=transfers`
-/// against an already-seeded cache. Shared by the per-call algorithm
-/// (which seeds a single τ) and the rRAPTOR scan (which re-seeds for
-/// each τ in descending order); the only difference between callers
-/// is the outer τ-loop that handles seeding.
-///
-/// **Caller invariants on entry:**
-///   - `marked_stops` contains the round-0 starting set: at minimum
-///     the sources, plus (for rRAPTOR) any stops where new trips just
-///     became catchable at the current τ.
-///   - `labels[0]` is seeded at every source with that source's
-///     departure label; the corresponding `best_arrival` and
-///     `ever_reached` bits are set.
-///   - `best_arrival` and `ever_reached` are coherent with `labels`
-///     (the helper reads and writes both).
-///
-/// On exit, `marked_stops` is empty (drained by the early-out check
-/// or by the natural marked-stops migration through rounds).
-#[allow(clippy::too_many_arguments)]
-fn run_raptor_rounds<T: Timetable + ?Sized, L: Label>(
-    tt: &T,
-    labels: &mut [Vec<LabelBag<L>>],
-    best_arrival: &mut [LabelBag<L>],
-    board_detail: &mut BoardingTree,
-    marked_stops: &mut FixedBitSet,
-    q_entry: &mut [Option<u32>],
-    q_routes: &mut Vec<RouteIdx>,
-    walked_buf: &mut Vec<StopIdx>,
-    relax_heap: &mut BinaryHeap<Reverse<(SecondOfDay, u32)>>,
-    ever_reached: &mut FixedBitSet,
-    transfers: usize,
-    targets: &[(StopIdx, Duration)],
-) {
-    let mut pt_threshold = best_to_any_target(best_arrival, targets);
-
-    // Pick the per-round footpath relaxation strategy once. Closed
-    // graphs use a single-pass O(E) walk; non-closed graphs need
-    // multi-source Dijkstra to chain walks to a fixed point.
-    let footpaths_closed = tt.footpaths_are_transitively_closed();
-
-    if footpaths_closed {
-        relax_footpaths_round_closed(
-            tt,
-            0,
-            labels,
-            best_arrival,
-            board_detail,
-            marked_stops,
-            pt_threshold,
-            walked_buf,
-            ever_reached,
-        );
-    } else {
-        relax_footpaths_round(
-            tt,
-            0,
-            labels,
-            best_arrival,
-            board_detail,
-            marked_stops,
-            pt_threshold,
-            walked_buf,
-            relax_heap,
-            ever_reached,
-        );
-    }
-    for s in walked_buf.drain(..) {
-        marked_stops.insert(s.idx());
-    }
-    pt_threshold = best_to_any_target(best_arrival, targets);
-
-    for k in 1..=transfers {
-        // Sparse carry-forward: only clone bags at stops ever
-        // reached so far this query. For real feeds the reached
-        // set is small (<10% of stops in typical city queries),
-        // so this avoids the 50k-stop dense memcpy that would
-        // otherwise dominate per-round overhead.
-        let (prev_labels, this_labels) = labels.split_at_mut(k);
-        let src = &prev_labels[k - 1];
-        let dst = &mut this_labels[0];
-        for bit in ever_reached.ones() {
-            dst[bit] = src[bit].clone();
-        }
-
-        // Build the route queue for this round. Each entry pairs a
-        // route with the earliest position on that route from which we
-        // can board this round. Stored positions are folded with `min`
-        // so multiple marked stops on the same route resolve to the
-        // earliest one.
-        for stop_bit in marked_stops.ones() {
-            let marked_stop = StopIdx::new(stop_bit as u32);
-            for &(route, pos) in tt.get_routes_serving_stop(marked_stop) {
-                match q_entry[route.idx()] {
-                    None => {
-                        q_entry[route.idx()] = Some(pos);
-                        q_routes.push(route);
-                    }
-                    Some(prev_pos) => {
-                        if pos < prev_pos {
-                            q_entry[route.idx()] = Some(pos);
-                        }
-                    }
-                }
-            }
-        }
-
-        marked_stops.clear();
-
-        // Per-route Pareto bag of riding entries. Each entry =
-        // (boarding_label, current_trip, boarding_stop); a single
-        // entry suffices for `ArrivalTime` (size-1 stop bags) but
-        // multi-criterion impls may carry multiple Pareto-optimal
-        // entries that boarded different trips.
-        let mut route_bag: SmallVec<[(L, TripIdx, StopIdx); 8]> = SmallVec::new();
-        let mut staged: SmallVec<[L; 8]> = SmallVec::new();
-
-        for &route in q_routes.iter() {
-            let p_pos = q_entry[route.idx()].expect("route in q_routes must have an entry");
-            route_bag.clear();
-
-            for (offset, &pi) in tt.get_stops_after(route, p_pos).iter().enumerate() {
-                let pos = p_pos + offset as u32;
-
-                // 1. Alight every active riding entry at pi.
-                for &(boarding_label, trip, boarding_stop) in route_bag.iter() {
-                    let arr = tt.get_arrival_time(trip, pos);
-                    let best_to_pi = best_arrival[pi.idx()].min_arrival();
-                    let time_to_beat = best_to_pi.min(pt_threshold);
-                    if arr >= time_to_beat {
-                        continue;
-                    }
-                    let new_label = boarding_label.extend_by_trip(arr);
-                    if labels[k][pi.idx()].insert(new_label) {
-                        best_arrival[pi.idx()].insert(new_label);
-                        board_detail.insert(
-                            (k, pi, new_label.arrival()),
-                            Step::Boarded {
-                                from: boarding_stop,
-                                route,
-                                parent_arrival: boarding_label.arrival(),
-                            },
-                        );
-                        marked_stops.insert(pi.idx());
-                        ever_reached.insert(pi.idx());
-                    }
-                }
-
-                // 2. Try to extend route_bag with labels from
-                //    labels[k-1][pi] that can catch a trip on this
-                //    route at pi. Snapshot first to avoid aliasing.
-                staged.clear();
-                staged.extend(labels[k - 1][pi.idx()].iter().copied());
-                for candidate in &staged {
-                    let cand_arr = candidate.arrival();
-                    let trip = match tt.get_earliest_trip(route, cand_arr, pos) {
-                        Some(t) => t,
-                        None => continue,
-                    };
-                    let trip_dep = tt.get_departure_time(trip, pos);
-
-                    // Redundancy check: existing route_bag entry
-                    // dominates candidate AND boards an at-or-earlier
-                    // trip at pi → candidate is redundant.
-                    let mut redundant = false;
-                    for &(l_existing, t_existing, _) in route_bag.iter() {
-                        let existing_dep = tt.get_departure_time(t_existing, pos);
-                        if l_existing.dominates(candidate) && existing_dep <= trip_dep {
-                            redundant = true;
-                            break;
-                        }
-                    }
-                    if redundant {
-                        continue;
-                    }
-
-                    // Remove entries strictly dominated by candidate.
-                    route_bag.retain(|&mut (l_existing, t_existing, _)| {
-                        let existing_dep = tt.get_departure_time(t_existing, pos);
-                        !(candidate.dominates(&l_existing) && trip_dep <= existing_dep)
-                    });
-                    route_bag.push((*candidate, trip, pi));
-                }
-            }
-        }
-
-        // Sparse-set reset of the route queue.
-        for r in q_routes.drain(..) {
-            q_entry[r.idx()] = None;
-        }
-
-        // Refresh target threshold after the route scan, then run
-        // footpath relax. Refresh again after the footpath round so
-        // the next round's boarding decisions use a current threshold.
-        pt_threshold = best_to_any_target(best_arrival, targets);
-        if footpaths_closed {
-            relax_footpaths_round_closed(
-                tt,
-                k,
-                labels,
-                best_arrival,
-                board_detail,
-                marked_stops,
-                pt_threshold,
-                walked_buf,
-                ever_reached,
-            );
-        } else {
-            relax_footpaths_round(
-                tt,
-                k,
-                labels,
-                best_arrival,
-                board_detail,
-                marked_stops,
-                pt_threshold,
-                walked_buf,
-                relax_heap,
-                ever_reached,
-            );
-        }
-        for s in walked_buf.drain(..) {
-            marked_stops.insert(s.idx());
-        }
-        pt_threshold = best_to_any_target(best_arrival, targets);
-
-        if marked_stops.is_clear() {
-            break;
-        }
-    }
-}
-
-/// Mark stops where any route through them has a trip whose departure
-/// time at that position falls in `[lo, hi)`. Used by rRAPTOR between
-/// τ scans to find stops that need rescanning because new trips just
-/// became catchable.
-///
-/// `marked` is the destination bitset; bits already set are preserved.
-/// One `get_earliest_trip` lookup per (route, position) plus one
-/// `get_departure_time` if the lookup hits — overall
-/// O(n_routes × max_route_len) calls per invocation, each O(log
-/// n_trips_per_route) inside the trait impl.
-fn newly_active_stops_into<T: Timetable + ?Sized>(
-    tt: &T,
-    lo: SecondOfDay,
-    hi: SecondOfDay,
-    marked: &mut FixedBitSet,
-) {
-    if lo >= hi {
-        return;
-    }
-    let n_routes = tt.n_routes() as u32;
-    for r in 0..n_routes {
-        let route = RouteIdx::new(r);
-        let stops = tt.get_stops_after(route, 0);
-        for (pos_offset, &stop) in stops.iter().enumerate() {
-            let pos = pos_offset as u32;
-            if let Some(trip) = tt.get_earliest_trip(route, lo, pos)
-                && tt.get_departure_time(trip, pos) < hi
-            {
-                marked.insert(stop.idx());
-            }
-        }
-    }
-}
 
 /// Models a route-based transit network for the RAPTOR algorithm.
 ///
@@ -1028,238 +380,8 @@ pub trait Timetable {
     where
         Self: Sized,
     {
-        let origins = origins.into_endpoints();
-        let targets = targets.into_endpoints();
-        let origins = origins.as_slice();
-        let targets = targets.as_slice();
-        cache.reset_for_query(transfers, self.n_stops() as u32, self.n_routes() as u32);
-        let RaptorCache {
-            labels,
-            best_arrival,
-            board_detail,
-            marked_stops,
-            q_entry,
-            q_routes,
-            walked_buf,
-            origin_set,
-            relax_heap,
-            ever_reached,
-            ..
-        } = cache;
-
-        // Clear and populate the origin set used by reconstruction.
-        origin_set.clear();
-        for &(o, _) in origins {
-            origin_set.insert(o.idx());
-        }
-
-        // Seed labels for each origin at depart + its walk-time offset.
-        // Reconstruction breaks the trace loop when it hits an origin
-        // (origin_set bit is set), so origins don't need a Step entry.
-        for &(o, walk) in origins {
-            let t = depart + walk;
-            let seed = L::from_departure(t);
-            if labels[0][o.idx()].insert(seed) {
-                best_arrival[o.idx()].insert(seed);
-                marked_stops.insert(o.idx());
-                ever_reached.insert(o.idx());
-            }
-        }
-
-        run_raptor_rounds(
-            self,
-            labels,
-            best_arrival,
-            board_detail,
-            marked_stops,
-            q_entry,
-            q_routes,
-            walked_buf,
-            relax_heap,
-            ever_reached,
-            transfers,
-            targets,
-        );
-
-        // For each target stop, enumerate every Pareto-optimal label
-        // in the target's bag at every round and reconstruct its plan.
-        // Effective label = bag label extended by the target's walk
-        // time offset.
-        let mut journeys =
-            extract_target_journeys(labels, board_detail, origin_set, targets, transfers);
-
-        // Output-side Pareto filter on (trip count, label). For any two
-        // returned journeys neither weakly dominates the other on the
-        // pair (plan.len, label). For single-criterion `ArrivalTime`
-        // this collapses to "strictly decreasing arrival as trip count
-        // increases" (the v0.10 contract); for multi-criterion impls
-        // it preserves Pareto-incomparable journeys (e.g. faster but
-        // more walking vs. slower but less walking).
-        //
-        // Sorted by plan.len ascending, then by `arrival()` ascending
-        // as a tiebreaker so the iteration order is deterministic.
-        journeys.sort_by_key(|j| (j.plan.len(), j.arrival()));
-        let mut front: Vec<Journey<L>> = Vec::with_capacity(journeys.len());
-        'outer: for j in journeys {
-            for f in &front {
-                if f.plan.len() <= j.plan.len() && f.label.dominates(&j.label) {
-                    continue 'outer;
-                }
-            }
-            front.retain(|f| !(j.plan.len() <= f.plan.len() && j.label.dominates(&f.label)));
-            front.push(j);
-        }
-        front
+        run_per_call_query(self, cache, transfers, depart, origins, targets)
     }
-}
-
-/// Pareto-profile filter for range-query output: keep entries that are
-/// not weakly dominated by another on `(later depart, fewer transfers,
-/// dominated label)`. Sort so the entries we'd prefer to keep come
-/// first (later departure, fewer transfers, earlier arrival), then
-/// sweep with mutual-domination removal.
-fn filter_range_pareto_front<L: Label>(mut all: Vec<RangeJourney<L>>) -> Vec<RangeJourney<L>> {
-    all.sort_by(|a, b| {
-        b.depart
-            .cmp(&a.depart)
-            .then(a.journey.plan.len().cmp(&b.journey.plan.len()))
-            .then(a.journey.arrival().cmp(&b.journey.arrival()))
-    });
-    let mut front: Vec<RangeJourney<L>> = Vec::with_capacity(all.len());
-    'outer: for r in all {
-        for f in &front {
-            if f.depart >= r.depart
-                && f.journey.plan.len() <= r.journey.plan.len()
-                && f.journey.label.dominates(&r.journey.label)
-            {
-                continue 'outer;
-            }
-        }
-        front.retain(|f| {
-            !(r.depart >= f.depart
-                && r.journey.plan.len() <= f.journey.plan.len()
-                && r.journey.label.dominates(&f.journey.label))
-        });
-        front.push(r);
-    }
-    front
-}
-
-/// Range-query algorithm specialised for `Label = ArrivalTime`. Single
-/// reverse-chronological scan that reuses labels across departure
-/// events (rRAPTOR, paper §4). For non-`ArrivalTime` labels, the
-/// parallel paths ([`Query::run_par`] / [`Query::run_with_pool`])
-/// still cover this case via the naïve batch.
-///
-/// `departures` must be sorted descending and deduped (the
-/// `.depart_in_window(...)` builder normalises this).
-///
-/// **Why descending τ order:** for τ' < τ_prev, a journey valid at
-/// τ_prev is also valid at τ' (the same trip with depart ≥ τ_prev ≥
-/// τ' is still catchable). So descending order makes each new seed
-/// monotonically improve the bag, and no improvement ever needs to
-/// be undone.
-///
-/// **Why labels accumulate (cache reset only once):** the per-call
-/// algorithm resets between queries; rRAPTOR resets only at the start
-/// of the whole scan. Labels written at scan τ remain valid for any
-/// τ' < τ, so persisting them across scans is correct, not stale —
-/// and avoids re-deriving information already discovered.
-///
-/// **Why `best_arrival` accumulates too:** the per-round carry-forward
-/// overwrites `labels[k][X]` at the start of each round k, but
-/// `best_arrival` is never cleared. This gives the `pt_threshold`
-/// pruning a tight bound across τ scans.
-fn raptor_range_rrap_arrival<T: Timetable + ?Sized>(
-    tt: &T,
-    cache: &mut RaptorCache<ArrivalTime>,
-    transfers: usize,
-    departures: &[SecondOfDay],
-    origins: Endpoints,
-    targets: Endpoints,
-) -> Vec<RangeJourney<ArrivalTime>> {
-    let origins = origins.as_slice();
-    let targets = targets.as_slice();
-
-    cache.reset_for_query(transfers, tt.n_stops() as u32, tt.n_routes() as u32);
-    let RaptorCache {
-        labels,
-        best_arrival,
-        board_detail,
-        marked_stops,
-        q_entry,
-        q_routes,
-        walked_buf,
-        origin_set,
-        relax_heap,
-        ever_reached,
-        ..
-    } = cache;
-
-    // origin_set is constant across τ scans; populate once.
-    origin_set.clear();
-    for &(source, _) in origins {
-        origin_set.insert(source.idx());
-    }
-
-    let mut output: Vec<RangeJourney<ArrivalTime>> = Vec::new();
-    let mut prev_tau: Option<SecondOfDay> = None;
-
-    for &tau in departures {
-        // (a) Seed sources at τ + walk. `LabelBag::insert` returns
-        // true iff the new label strictly improves the bag; for τ <
-        // prev_tau the new label dominates the prior one, so
-        // re-seeding tightens the source bag. For the first scan
-        // there's nothing to dominate and the seed is added directly.
-        marked_stops.clear();
-        for &(source, walk) in origins {
-            let seed = ArrivalTime(tau + walk);
-            if labels[0][source.idx()].insert(seed) {
-                best_arrival[source.idx()].insert(seed);
-                marked_stops.insert(source.idx());
-                ever_reached.insert(source.idx());
-            }
-        }
-
-        // (b) Mark stops with newly-catchable trips for this τ. The
-        // window is half-open `[tau, prev_tau)` — trips departing at
-        // exactly `tau` are first-catchable in this scan, while trips
-        // at `prev_tau` were already covered by the previous scan.
-        if let Some(prev) = prev_tau {
-            newly_active_stops_into(tt, tau, prev, marked_stops);
-        }
-
-        // (c) Round-0 footpath relax + rounds 1..=transfers, sharing
-        // labels with previous scans.
-        run_raptor_rounds(
-            tt,
-            labels,
-            best_arrival,
-            board_detail,
-            marked_stops,
-            q_entry,
-            q_routes,
-            walked_buf,
-            relax_heap,
-            ever_reached,
-            transfers,
-            targets,
-        );
-
-        // (d) Snapshot per-target journeys for this τ.
-        let snapshot =
-            extract_target_journeys(labels, board_detail, origin_set, targets, transfers);
-        for journey in snapshot {
-            output.push(RangeJourney {
-                depart: tau,
-                journey,
-            });
-        }
-
-        prev_tau = Some(tau);
-    }
-
-    filter_range_pareto_front(output)
 }
 
 /// One entry in a range-query profile: a departure time paired with
@@ -1560,44 +682,44 @@ where
 /// `RaptorCache` is `!Sync` — give each worker its own, or use
 /// [`RaptorCachePool`] (the `Sync` freelist variant).
 pub struct RaptorCache<L: Label = ArrivalTime> {
-    n_stops: u32,
-    n_routes: u32,
+    pub(crate) n_stops: u32,
+    pub(crate) n_routes: u32,
 
     /// labels[k][stop.idx()] = Pareto bag of labels at stop with at
     /// most k trips. Empty bag means unreached.
-    labels: Vec<Vec<LabelBag<L>>>,
+    pub(crate) labels: Vec<Vec<LabelBag<L>>>,
 
     /// τ* — Pareto bag of best labels at each stop across all rounds.
-    best_arrival: Vec<LabelBag<L>>,
+    pub(crate) best_arrival: Vec<LabelBag<L>>,
 
     /// Boarding tree for journey reconstruction.
-    board_detail: BoardingTree,
+    pub(crate) board_detail: BoardingTree,
 
     /// Bitset of marked stops, sized to n_stops.
-    marked_stops: FixedBitSet,
+    pub(crate) marked_stops: FixedBitSet,
 
     /// Per-round route queue. `q_entry[r.idx()] = Some(boarding_pos)` when
     /// route r has been entered this round; `q_routes` is the dense list
     /// of routes that have entries (for cheap iteration without scanning
     /// `n_routes` empty slots).
-    q_entry: Vec<Option<u32>>,
-    q_routes: Vec<RouteIdx>,
+    pub(crate) q_entry: Vec<Option<u32>>,
+    pub(crate) q_routes: Vec<RouteIdx>,
 
     /// Scratch buffer for footpath relaxation output.
-    walked_buf: Vec<StopIdx>,
+    pub(crate) walked_buf: Vec<StopIdx>,
 
     /// Bitset of origin stops for the current query. Reconstruction
     /// terminates when the trace reaches any bit set here.
-    origin_set: FixedBitSet,
+    pub(crate) origin_set: FixedBitSet,
 
     /// Min-heap reused across footpath relaxations. Entries are
     /// `(arrival_time, stop_bit)`; lazy deletion via the time field.
-    relax_heap: BinaryHeap<Reverse<(SecondOfDay, u32)>>,
+    pub(crate) relax_heap: BinaryHeap<Reverse<(SecondOfDay, u32)>>,
 
     /// Bitset of stops with a non-empty bag in any round seen so
     /// far this query. Used to make per-round carry-forward sparse:
     /// only clone bags at set bits, not all `n_stops`.
-    ever_reached: FixedBitSet,
+    pub(crate) ever_reached: FixedBitSet,
 }
 
 impl<L: Label> RaptorCache<L> {
@@ -1626,7 +748,7 @@ impl<L: Label> RaptorCache<L> {
         }
     }
 
-    fn reset_for_query(&mut self, transfers: K, tt_n_stops: u32, tt_n_routes: u32) {
+    pub(crate) fn reset_for_query(&mut self, transfers: K, tt_n_stops: u32, tt_n_routes: u32) {
         assert_eq!(
             self.n_stops, tt_n_stops,
             "RaptorCache sized for {} stops but timetable has {}",
